+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to eww will be listed here, starting at changes since versio
 - Made `and`, `or` and `?:` lazily evaluated in simplexpr (By: ModProg)
 - Add Vanilla CSS support (By: Ezequiel Ramis)
 - Add `jq` function, offering jq-style json processing
+- Add `justify` property to the label widget, allowing text justification (By: n3oney)
 
 ## [0.4.0] (04.09.2022)